requirements.txt
yandex_tracker_client>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
click>=8.0.0